Restructure tests for maintainability and ensure consecutive modified lines are trimmed. Fixes #43. (#44)
* Restructure tests to use per-test sample and result files.

The intention is to make it easier to modify and extend individual tests
without impacting the entire suite. The repeated file loading does slow
down the test suite by a factor of about 5, but it's a reasonable small
suite so the real impact of this is minimal.

* Modified test assert to propagate assert errors through done().

Previously, the test would fail with an unfulfilled promise, with the
error description given on stdout. Now the assertion error is caught and
passed to done, which gives a nicer output.

* Added tests to for trimming consecutive modified lines. Tests #43.

At time of commit, these tests fail - only the first line of any block
of consecutive modified lines is trimmed.

* Address failure to trim all consecutive modified lines. Fixes #43.